+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+require(__DIR__.'/functionOverwrites.php');
+
+require(__DIR__.'/../src/symmetricEncryption.php');